split actions
