+++ src/actions/index.js
@@ -1,9 +1,15 @@
 import * as app from './app';
 import * as fonts from './fonts';
 import * as words from './words';
+import * as logo from './logo';
+import * as suffix from './suffix';
+import * as tagline from './tagline';
 
 export {
   app,
   fonts,
-  words
+  words,
+  logo,
+  suffix,
+  tagline
 };
