+++ src/actions/suffix.js
@@ -1,0 +1,7 @@
+/**
+ * Updates state.suffix
+ * @param {string} suffix
+ */
+export function set(store, suffix) {
+  store.setState({ suffix });
+}
